Debug: mobile dark mode debug (#373)
* feat: dark mode & new messages on welcome modal

* fix: dark mode on sign in modal & user login icon
