+++ syllabus/package.json
@@ -88,7 +88,7 @@
     "@bit/wasedatime.core.ts.ui.loading-spinner": "0.1.2",
     "@bit/wasedatime.core.ts.ui.logo": "0.0.9",
     "@bit/wasedatime.core.ts.ui.modal": "0.0.8",
-    "@bit/wasedatime.core.ts.ui.sign-in-modal": "npm:wasedatime.core.ts.ui.sign-in-modal@0.1.4",
+    "@bit/wasedatime.core.ts.ui.sign-in-modal": "0.1.6",
     "@bit/wasedatime.core.ts.utils.i18n": "0.0.4",
     "@bit/wasedatime.core.ts.utils.responsive-utils": "0.0.1",
     "@bit/wasedatime.core.ts.utils.user": "npm:wasedatime.core.ts.utils.user@0.0.4",
